prod: aws fix for static-nodes.json and enode.
* when generating the enode it needs to be save to a file
so that the update_enodes_configmap function can access it.
* when using 'replace' with kubectl the configmap needs to exist,
this was failing for initial generationfor goquorum-node-enodes.
the fix is to create an empty configmap if one does not exist, and if
one does not exist the static-nodes.json.raw should only contain []
until the enode url is added. The previous behavior would error if the
goquorum-node-enodes and overwrite the "[]" in static-nodes.json.raw
causing errors.
* adding intermediate variable STATIC_NODES_EXISTS and additional echo
statements for clarity and future debuggers.
* note: kubectl -n {{ .Release.Namespace }} create configmap goquorum-node-enodes
--from-file=static-nodes.json=/tmp/static-nodes.json -o yaml
--dry-run=client | kubectl replace -f -

apply doesn't work as the permission in aws do not allow "patch"
permissions.

"system:serviceaccount:quorum:quorum-node-secrets-sa" cannot patch resource "configmaps"

see:
  https://github.com/kubernetes/kubernetes/issues/65066
  https://github.com/kubernetes/kubernetes/issues/89129